add setup to simplify
