+++ src/components/Customs/Selects/SyAutocomplete/SyAutocomplete.vue
@@ -4,35 +4,18 @@
 	})
 
 	import { mdiAlertCircle, mdiChevronDown, mdiCloseCircle } from '@mdi/js'
-	import { computed, nextTick, onMounted, ref, watch, watchEffect, type PropType } from 'vue'
-
-	import { useSySelectKeyboard } from '../SySelect/composables/useSySelectKeyboard'
+	import { type PropType } from 'vue'
 	import { vRgaaSvgFix } from '../../../../directives/rgaaSvgFix'
-	import { useValidatable } from '@/composables/validation/useValidatable'
-	import type { VList as VListComponent } from 'vuetify/components'
 	import { VChip, VList, VTextField, VListItem, VListItemTitle, VMenu } from 'vuetify/components'
 	import SyCheckbox from '@/components/Customs/SyCheckbox/SyCheckbox.vue'
 	import SyIcon from '@/components/Customs/SyIcon/SyIcon.vue'
-	import { sanitizeHtml } from '@/utils/sanitizeHtml'
 
 	import { locales } from './locales'
 
-	import type { ItemType, SelectItemArrayType, SelectItemValueType } from './types'
-	import { useSyAutocompleteFetch } from './composables/useSyAutocompleteFetch'
-	import { useSyAutocompleteValidation } from './composables/useSyAutocompleteValidation'
-	import { useSyAutocompleteAria } from './composables/useSyAutocompleteAria'
-	import { useSyAutocompleteVuetifyAdapter } from './composables/useSyAutocompleteVuetifyAdapter'
-	import { useSyAutocompleteModel } from './composables/useSyAutocompleteModel'
-	import { useSyAutocompleteMenu } from './composables/useSyAutocompleteMenu'
-	import { useSyAutocompleteSelection } from './composables/useSyAutocompleteSelection'
-	import { useSyAutocompleteKeyboardOpen } from './composables/useSyAutocompleteKeyboardOpen'
-	import { useSyAutocompleteKeydown } from './composables/useSyAutocompleteKeydown'
-	import { useSyComboboxFormatItems } from '../common/combobox/useSyComboboxFormatItems'
-	import { useSyComboboxIds } from '../common/combobox/useSyComboboxIds'
-	import { useSyComboboxCalculatedWidth } from '../common/combobox/useSyComboboxCalculatedWidth'
-	import { useSyComboboxMenuTarget } from '../common/combobox/useSyComboboxMenuTarget'
+	import type { ItemType, SelectItemArrayType } from './types'
+	import { useSyAutocompleteSetup } from './composables/useSyAutocompleteSetup'
 
-	export type { ItemType, SelectItemArrayType, SelectItemValueType } from './types'
+	export type { ItemType, SelectItemArrayType } from './types'
 
 	const props = defineProps({
 		modelValue: {
@@ -169,422 +152,47 @@
 		'loading',
 	])
 
-	const isOpen = ref(false)
-	const { inputId, uniqueMenuId } = useSyComboboxIds({
-		inputIdPrefix: 'sy-autocomplete',
-		defaultMenuId: 'sy-autocomplete-menu',
-		menuId: computed(() => props.menuId),
-	})
-	const optionIdPrefix = computed(() => `${uniqueMenuId.value}-`)
-
-	// Valeur sélectionnée (single) ou tableau (multiple). Elle sert de source de vérité interne.
-	const selectedItem = ref<SelectItemValueType | SelectItemArrayType>(props.modelValue)
-	// Valeur tapée dans l'input (v-model:search). Peut être différente de selectedItem.
-	const searchValue = ref(props.search)
-
-	// Normalisation: props.errorMessages peut être string | string[].
-	// On uniformise en tableau pour simplifier les composables de validation.
-	const normalizedErrorMessages = computed<readonly string[]>(() => {
-		if (typeof props.errorMessages === 'string') {
-			return props.errorMessages ? [props.errorMessages] : []
-		}
-		return props.errorMessages ?? []
-	})
-
-	const emitLoading = (value: boolean) => {
-		emit('loading', value)
-	}
-	const emitError = (error: unknown) => {
-		emit('error', error)
-	}
-
 	const {
-		internalItems,
+		isOpen,
+		inputId,
+		uniqueMenuId,
+		optionIdPrefix,
+		selectedItem,
+		textFieldModel,
 		isLoading,
-		scheduleFetch,
-		resetFetchState,
-		syncItemsFromProps,
-	} = useSyAutocompleteFetch({
-		// Centralise: filtrage local ou appel API + debounce + cache + "last request wins".
-		items: computed(() => props.items),
-		fetchItems: props.fetchItems,
-		minChars: computed(() => props.minChars),
-		debounceMs: computed(() => props.debounceMs),
-		cache: computed(() => props.cache),
-		getItemText: (item: unknown) => (item as Record<string, unknown>)[props.textKey] as string,
-		emitLoading,
-		emitError,
-	})
-
-	const {
+		isNoDataVisible,
 		hasError,
-		isTouched,
-		isRequired,
-		computedHasError,
 		textFieldErrorMessages,
 		requiredRules,
-		markTouched,
+		showHelpTextAsMessage,
+		showHelpTextBelow,
 		validateOnSubmit,
-	} = useSyAutocompleteValidation({
-		// Centralise: notion de "touched" + affichage des erreurs + règle required.
-		required: computed(() => props.required),
-		errorMessages: normalizedErrorMessages,
-		readonly: computed(() => props.readonly),
-		disableErrorHandling: computed(() => props.disableErrorHandling),
-		multiple: computed(() => props.multiple),
-		selectedItem,
-	})
-
-	const labelWidth = ref(0)
-	const labelRef = ref<HTMLElement | null>(null)
-	const list = ref<VListComponent | null>(null)
-	const textInput = ref<InstanceType<typeof VTextField> | null>(null)
-	const htmlItemRefs = ref<HTMLElement[]>([])
-	const isInputFocused = ref(false)
-	const openedByTyping = ref(false)
-	const pendingFocusIndex = ref<number | null>(null)
-
-	const { menuTarget } = useSyComboboxMenuTarget({
-		textInput: textInput as unknown as import('vue').Ref<{ $el?: HTMLElement } | null>,
-	})
-
-	const { calculatedWidth } = useSyComboboxCalculatedWidth({
-		width: computed(() => props.width),
-	})
-
-	const isShouldDisplayAsterisk = computed(() => props.required && props.displayAsterisk)
-	const labelWithAsterisk = computed(() => isShouldDisplayAsterisk.value ? `${props.label} *` : props.label)
-
-	const hasSelectionToClear = computed(() => {
-		return props.multiple
-			? (((selectedItem.value as unknown[] | null | undefined)?.length) ?? 0) > 0
-			: selectedItem.value != null
-	})
-
-	const hasMessages = computed(() => {
-		if (props.disableErrorHandling) return false
-		return normalizedErrorMessages.value.length > 0 || hasError.value
-	})
-
-	const showHelpTextAsMessage = computed(() => {
-		return props.helpText && !hasMessages.value
-	})
-
-	const showHelpTextBelow = computed(() => {
-		return props.helpText && hasMessages.value && !props.hideMessages
-	})
-
-	const updateHasErrorFromSelection = () => {
-		hasError.value = computedHasError.value
-	}
-
-	const textFieldModel = computed({
-		get: () => {
-			if (props.multiple && !props.chips && searchValue.value.trim().length === 0) {
-				return getMultipleSelectionText()
-			}
-			return searchValue.value
-		},
-		set: (value: string) => {
-			searchValue.value = value
-		},
-	})
-
-	let focusInputElement = () => {}
-	let ensureNativeInputFocus = () => {}
-	let setActiveDescendantForMenu = (index: number) => {
-		void index
-	}
-
-	const handleInputFocus = () => {
-		isInputFocused.value = true
-		ensureNativeInputFocus()
-	}
-
-	const handleInputBlur = () => {
-		isInputFocused.value = false
-		markTouched()
-		if (props.multiple && !props.chips) {
-			// À la sortie du champ, on affiche les valeurs sélectionnées plutôt qu'une recherche transitoire.
-			searchValue.value = ''
-		}
-	}
-
-	const {
-		openMenu,
-		toggleMenu,
-		closeList,
-		markOpenedByTyping,
-	} = useSyAutocompleteMenu({
-		readonly: computed(() => props.readonly),
-		multiple: computed(() => props.multiple),
-		isOpen,
+		labelWithAsterisk,
+		labelWidth,
+		labelRef,
 		list,
-		searchValue,
-		ensureNativeInputFocus: () => ensureNativeInputFocus(),
-		scheduleFetch,
-		setActiveDescendant: index => setActiveDescendantForMenu(index),
-		openedByTyping,
-	})
-
-	const {
+		textInput,
+		menuTarget,
+		calculatedWidth,
+		htmlItemRefs,
 		hasChips,
-		getItemText,
-		isItemSelected,
-		getPlainItemText,
+		hasSelectionToClear,
+		clearSelection,
+		removeChip,
 		getChipText,
 		getChipKey,
-		getMultipleSelectionText,
-		removeChip,
-		clearSelection,
-		selectItem,
-	} = useSyAutocompleteSelection({
-		multiple: computed(() => props.multiple),
-		chips: computed(() => props.chips),
-		returnObject: computed(() => props.returnObject),
-		textKey: computed(() => props.textKey),
-		valueKey: computed(() => props.valueKey),
-		plainTextKey: computed(() => props.plainTextKey),
-		allowHtml: computed(() => props.allowHtml),
-		internalItems,
-		selectedItem,
-		searchValue,
-		isOpen,
-		markTouched,
-		updateHasError: updateHasErrorFromSelection,
-		ensureNativeInputFocus: () => ensureNativeInputFocus(),
-		emitUpdateModelValue: value => emit('update:modelValue', value),
-		emitUpdateSearch: value => emit('update:search', value),
-	})
-
-	const getSelectedText = () => {
-		if (props.multiple) return null
-		if (selectedItem.value == null) return null
-
-		if (props.returnObject) {
-			return String(getPlainItemText(selectedItem.value as unknown) ?? '')
-		}
-
-		const selectedValue = selectedItem.value
-		const match = internalItems.value.find((it) => {
-			if (!it || typeof it !== 'object') return false
-			return (it as Record<string, unknown>)[props.valueKey] === selectedValue
-		})
-		return match ? String(getPlainItemText(match as unknown) ?? '') : ''
-	}
-
-	const { formattedItems } = useSyComboboxFormatItems({
-		items: computed(() => internalItems.value),
-		textKey: computed(() => props.textKey),
-		valueKey: computed(() => props.valueKey),
-	})
-
-	// Gestion clavier (navigation liste / sélection / escape / home/end/page up/down).
-	// Important: garde le focus DOM sur l'input (pattern combobox) et utilise aria-activedescendant.
-	const {
-		activeDescendantId,
-		setActiveDescendant,
-		clearActiveDescendant,
-		handleEnterKey,
-		handleSpaceKey,
-		handleDownKey,
-		handleUpKey,
-		handleEscapeKey,
-		handleHomeKey,
-		handleEndKey,
-		handlePageUpKey,
-		handlePageDownKey,
-		handleTabKey,
-	} = useSySelectKeyboard({
-		isOpen,
 		formattedItems,
-		toggleMenu,
-		selectItem,
 		getItemText,
-		optionIdPrefix,
-		focusOptions: false,
-		restoreOnOpen: false,
-		initialFocusIndex: 0,
-	})
-
-	setActiveDescendantForMenu = setActiveDescendant
-
-	// Gestion ARIA (attributs combobox/listbox) + mise à jour des états (open/activeDescendant/error/required).
-	useSyAutocompleteAria({
-		textInput,
-		isOpen,
-		uniqueMenuId,
+		isItemSelected,
+		selectItem,
 		activeDescendantId,
-		isRequired,
-		hasError,
-		selectedItem,
-	})
-
-	const isNoDataVisible = computed(() => {
-		return !isLoading.value
-			&& searchValue.value.trim().length >= props.minChars
-			&& formattedItems.value.length === 0
-	})
-
-	watchEffect(() => {
-		if (!props.allowHtml) {
-			return
-		}
-		htmlItemRefs.value.forEach((el, index) => {
-			const item = formattedItems.value[index]
-			if (!el || !item) {
-				return
-			}
-			el.innerHTML = sanitizeHtml(String(getItemText(item) ?? ''))
-		})
-	})
-
-	useSyAutocompleteModel({
-		modelValue: computed(() => props.modelValue as SelectItemValueType | SelectItemArrayType),
-		search: computed(() => props.search),
-		multiple: computed(() => props.multiple),
-		minChars: computed(() => props.minChars),
-		getSelectedText,
-
-		selectedItem,
-		searchValue,
-		isOpen,
-		pendingFocusIndex,
-		openMenu,
-		resetFetchState,
-		scheduleFetch,
-		clearActiveDescendant,
-		markTouched,
-		markOpenedByTyping,
-		emitUpdateModelValue: value => emit('update:modelValue', value),
-		emitUpdateSearch: value => emit('update:search', value),
-	})
-
-	watch(() => props.items, (newItems) => {
-		syncItemsFromProps(newItems)
-	})
-
-	watch([selectedItem, isRequired, isTouched], () => {
-		if (props.disableErrorHandling || props.readonly) {
-			hasError.value = false
-			return
-		}
-
-		if (isTouched.value) {
-			hasError.value = computedHasError.value
-		}
-	})
-
-	watch([hasError], ([newHasError]) => {
-		if (props.disableErrorHandling || props.readonly) {
-			hasError.value = false
-			return
-		}
-
-		hasError.value = newHasError
-	})
-
-	useValidatable(validateOnSubmit)
-
-	onMounted(() => {
-		if (labelRef.value) {
-			labelWidth.value = labelRef.value.offsetWidth + 64
-		}
-	})
-
-	const ensureFirstOptionFocused = () => {
-		setActiveDescendant(0)
-	}
-
-	const handleListEscapeKey = () => {
-		closeList()
-		nextTick(() => {
-			focusInputElement()
-		})
-	}
-
-	const {
-		handleInputDownKey,
-		handleInputUpKey,
-		handleListDownKey,
-		handleListUpKey,
-	} = useSyAutocompleteKeyboardOpen({
-		isOpen,
-		activeDescendantId,
-		optionIdPrefix,
-		formattedItemsLength: computed(() => formattedItems.value.length),
-		pendingFocusIndex,
-		openMenu,
-		clearActiveDescendant,
-		setActiveDescendant,
-		handleDownKey,
-		handleUpKey,
-		ensureFirstOptionFocused,
-	})
-
-	const { onNativeInputKeydown, onFieldRootKeydown, onListKeydown } = useSyAutocompleteKeydown({
-		textInput,
-		isOpen,
-		multiple: computed(() => props.multiple),
-		selectedItem,
-		searchValue,
-		handleTabKey,
-		handleInputDownKey,
-		handleInputUpKey,
-		handleListDownKey,
-		handleListUpKey,
-		handleListEscapeKey,
-		handleHomeKey,
-		handleEndKey,
-		handlePageUpKey,
-		handlePageDownKey,
-		handleEnterKey,
-		handleSpaceKey,
-		handleEscapeKey,
-		removeChip,
-		focusInputElement: () => focusInputElement(),
-		ensureNativeInputFocus: () => ensureNativeInputFocus(),
-	})
-
-	// Adaptateur Vuetify/DOM : gestion du focus natif et des listeners clavier sur les bons éléments.
-	;({ focusInputElement, ensureNativeInputFocus } = useSyAutocompleteVuetifyAdapter({
-		textInput,
-		list,
-		isOpen,
-		activeDescendantId,
-		onNativeInputKeydown,
-		onFieldRootKeydown,
-	}))
-
-	watch(isOpen, (newValue) => {
-		if (!newValue) {
-			openedByTyping.value = false
-		}
-		if (newValue) {
-			nextTick(() => {
-				// On s'assure d'avoir une option active à l'ouverture (comportement type SySelect)
-				if (!openedByTyping.value && !activeDescendantId.value) {
-					setActiveDescendant(0)
-				}
-				openedByTyping.value = false
-			})
-		}
-	})
-
-	type ActivatorProps = Record<string, unknown> & {
-		onClick?: unknown
-		ref?: (el: unknown) => void
-	}
-	function initializeActivatorProps(activatorProps: ActivatorProps) {
-		return {
-			...activatorProps,
-			onClick: undefined,
-			ref: (el) => {
-				textInput.value = el as InstanceType<typeof VTextField>
-				activatorProps.ref?.(el)
-			},
-		}
-	}
+		closeList,
+		toggleMenu,
+		handleInputFocus,
+		handleInputBlur,
+		onListKeydown,
+		initializeActivatorProps,
+	} = useSyAutocompleteSetup(props as unknown as Parameters<typeof useSyAutocompleteSetup>[0], emit)
 
 	defineExpose({
 		validateOnSubmit,
