+++ src/components/Customs/Selects/SySelect/composables/useSySelectSetup.ts
@@ -1,0 +1,331 @@
+import { computed, onMounted, ref, watch, watchEffect, type ComputedRef, type Ref } from 'vue'
+import type { VList, VTextField } from 'vuetify/components'
+
+import { useSySelectAria } from './useSySelectAria'
+import { useSySelectKeyboard } from './useSySelectKeyboard'
+import { useSySelectKeydown } from './useSySelectKeydown'
+import { useSySelectMenu } from './useSySelectMenu'
+import { useSySelectSelection } from './useSySelectSelection'
+import { useSySelectValidation } from './useSySelectValidation'
+import { useSySelectVuetifyAdapter } from './useSySelectVuetifyAdapter'
+import { useSyComboboxCalculatedWidth } from '../../common/combobox/useSyComboboxCalculatedWidth'
+import { useSyComboboxFormatItems } from '../../common/combobox/useSyComboboxFormatItems'
+import { useSyComboboxIds } from '../../common/combobox/useSyComboboxIds'
+import { useSyComboboxMenuTarget } from '../../common/combobox/useSyComboboxMenuTarget'
+import { sanitizeHtml } from '@/utils/sanitizeHtml'
+
+export type SySelectSetupItemType = {
+	[key: string]: unknown
+}
+
+export type SySelectSetupItemValueType = Record<string, unknown> | string | number | null | undefined
+export type SySelectSetupItemArrayType = Array<Record<string, unknown> | string | number>
+
+export type SySelectSetupEmit = (event: 'update:modelValue', value: unknown) => void
+
+export type SySelectSetupProps = {
+	modelValue: SySelectSetupItemValueType | SySelectSetupItemArrayType
+	items: SySelectSetupItemType[]
+	label: string
+	errorMessages: string | readonly string[]
+	required: boolean
+	disabled: boolean
+	menuId: string
+	outlined: boolean
+	textKey: string
+	plainTextKey: string
+	valueKey: string
+	displayAsterisk: boolean
+	returnObject: boolean
+	disableErrorHandling: boolean
+	density: 'default' | 'comfortable' | 'compact' | undefined
+	bgColor: string
+	readonly: boolean
+	clearable: boolean
+	hideMessages: boolean
+	width: string
+	multiple: boolean
+	chips: boolean
+	helpText: string
+	allowHtml: boolean
+	autocomplete: string
+}
+
+export type SySelectSetupResult = {
+	isOpen: Ref<boolean>
+	labelWidth: Ref<number>
+	inputId: Ref<string>
+	uniqueMenuId: Ref<string>
+	formattedItems: ComputedRef<SySelectSetupItemType[]>
+	hasError: Ref<boolean>
+	isRequired: ComputedRef<boolean>
+	showHelpTextAsMessage: ComputedRef<boolean | ''>
+	showHelpTextBelow: ComputedRef<boolean | ''>
+	validateOnSubmit: () => void
+	labelWithAsterisk: ComputedRef<string>
+	calculatedWidth: ComputedRef<string | undefined>
+	menuTarget: ComputedRef<HTMLElement | undefined>
+	labelRef: Ref<HTMLElement | null>
+	htmlItemRefs: Ref<HTMLElement[]>
+	textInput: Ref<InstanceType<typeof VTextField> | null>
+	list: Ref<VList | null>
+	toggleMenu: () => void
+	closeList: () => void
+	activeDescendantId: Ref<string>
+	onFieldKeydown: (event: KeyboardEvent) => void
+	onListKeydown: (event: KeyboardEvent) => void
+	isDefaultOption: (item: SySelectSetupItemType) => boolean
+	isItemSelected: (item: SySelectSetupItemType) => boolean
+	selectItem: (item: SySelectSetupItemType | null, event?: Event) => void
+	removeChip: (item: SySelectSetupItemType | string | number) => void
+	getItemText: (item: SySelectSetupItemType | string | number) => string
+	getChipText: (item: SySelectSetupItemType | string | number) => string
+	hasChips: Ref<boolean>
+	selectedChipsItems: ComputedRef<Array<SySelectSetupItemType | string | number>>
+	hasSelectionToClear: ComputedRef<boolean>
+	selectedItemText: ComputedRef<string>
+	initializeActivatorProps: (activatorProps: Record<string, unknown>) => Record<string, unknown>
+}
+
+export function useSySelectSetup(props: SySelectSetupProps, emit: SySelectSetupEmit): SySelectSetupResult {
+	const isOpen = ref(false)
+	const selectedItem = ref<SySelectSetupItemValueType | SySelectSetupItemArrayType>(props.modelValue)
+
+	const { hasError, isRequired, showHelpTextAsMessage, showHelpTextBelow, validateOnSubmit } = useSySelectValidation({
+		isOpen,
+		selectedItem,
+		disableErrorHandling: computed(() => props.disableErrorHandling),
+		readonly: computed(() => props.readonly),
+		required: computed(() => props.required),
+		errorMessages: computed(() => (Array.isArray(props.errorMessages) ? props.errorMessages : [props.errorMessages])),
+		helpText: computed(() => props.helpText),
+		hideMessages: computed(() => props.hideMessages),
+	})
+
+	const labelWidth = ref(0)
+	const labelRef = ref<HTMLElement | null>(null)
+	const list = ref<VList | null>(null)
+	const textInput = ref<InstanceType<typeof VTextField> | null>(null)
+	const htmlItemRefs = ref<HTMLElement[]>([])
+
+	const vuetifyTextInputAdapter = computed(() => {
+		return textInput.value
+			? { $el: textInput.value.$el as HTMLElement }
+			: null
+	})
+
+	const { ensureNativeInputFocus } = useSySelectVuetifyAdapter({
+		textInput: vuetifyTextInputAdapter,
+	})
+
+	const { inputId, uniqueMenuId } = useSyComboboxIds({
+		inputIdPrefix: 'sy-select',
+		defaultMenuId: 'sy-select-menu',
+		menuId: computed(() => props.menuId),
+	})
+
+	const { formattedItems } = useSyComboboxFormatItems({
+		items: computed(() => props.items),
+		textKey: computed(() => props.textKey),
+		valueKey: computed(() => props.valueKey),
+	})
+
+	let setActiveDescendantForMenu = (index: number) => {
+		void index
+	}
+	let isItemSelectedForMenu = (item: SySelectSetupItemType) => {
+		void item
+		return false
+	}
+
+	const { toggleMenu, closeList } = useSySelectMenu({
+		readonly: computed(() => props.readonly),
+		multiple: computed(() => props.multiple),
+		isOpen,
+		list: list as unknown as Ref<{ $el: HTMLElement } | null>,
+		formattedItems,
+		isItemSelected: item => isItemSelectedForMenu(item),
+		setActiveDescendant: index => setActiveDescendantForMenu(index),
+	})
+
+	let setActiveDescendantForSelection = (index: number) => {
+		void index
+	}
+	let restoreFocusForSelection = () => {}
+
+	const {
+		isDefaultOption,
+		isItemSelected,
+		selectItem,
+		removeChip,
+		getItemText,
+		safeChipItem,
+		getChipText,
+		hasChips,
+		selectedChipsItems,
+		hasSelectionToClear,
+		selectedItemText,
+	} = useSySelectSelection({
+		items: computed(() => props.items),
+		formattedItems,
+		selectedItem,
+		multiple: computed(() => props.multiple),
+		chips: computed(() => props.chips),
+		returnObject: computed(() => props.returnObject),
+		textKey: computed(() => props.textKey),
+		plainTextKey: computed(() => props.plainTextKey),
+		valueKey: computed(() => props.valueKey),
+		allowHtml: computed(() => props.allowHtml),
+		isOpen,
+		ensureNativeInputFocus: () => ensureNativeInputFocus(),
+		setActiveDescendant: index => setActiveDescendantForSelection(index),
+		restoreFocus: () => restoreFocusForSelection(),
+		emitUpdateModelValue: value => emit('update:modelValue', value),
+	})
+	void safeChipItem
+	isItemSelectedForMenu = (item: SySelectSetupItemType) => isItemSelected(item)
+
+	watchEffect(() => {
+		if (!props.allowHtml) {
+			return
+		}
+
+		htmlItemRefs.value.forEach((el, index) => {
+			const item = formattedItems.value[index]
+			if (!el || !item) {
+				return
+			}
+			el.innerHTML = sanitizeHtml(String(getItemText(item) ?? ''))
+		})
+	})
+
+	const labelWithAsterisk = computed(() => {
+		return props.required && props.displayAsterisk ? `${props.label} *` : props.label
+	})
+
+	const { calculatedWidth } = useSyComboboxCalculatedWidth({
+		width: computed(() => props.width),
+	})
+
+	const { menuTarget } = useSyComboboxMenuTarget({
+		textInput: textInput as unknown as Ref<{ $el?: HTMLElement } | null>,
+	})
+
+	watch(() => props.modelValue, (newValue) => {
+		selectedItem.value = newValue
+	})
+
+	const {
+		activeDescendantId,
+		setActiveDescendant,
+		handleEnterKey,
+		handleSpaceKey,
+		handleDownKey,
+		handleUpKey,
+		handleCharacterKey,
+		handleEscapeKey,
+		handleHomeKey,
+		handleEndKey,
+		handlePageUpKey,
+		handlePageDownKey,
+		handleTabKey,
+		restoreFocus,
+	} = useSySelectKeyboard({
+		isOpen,
+		formattedItems,
+		toggleMenu,
+		selectItem,
+		getItemText,
+	})
+
+	setActiveDescendantForMenu = (index: number) => {
+		setActiveDescendant(index)
+	}
+	setActiveDescendantForSelection = (index: number) => {
+		setActiveDescendant(index)
+	}
+	restoreFocusForSelection = () => {
+		restoreFocus()
+	}
+
+	const { onFieldKeydown, onListKeydown } = useSySelectKeydown({
+		handleEnterKey,
+		handleSpaceKey,
+		handleDownKey,
+		handleUpKey,
+		handleEscapeKey,
+		handleHomeKey,
+		handleEndKey,
+		handlePageUpKey,
+		handlePageDownKey,
+		handleTabKey,
+		handleCharacterKey,
+		closeList,
+	})
+
+	onMounted(() => {
+		if (labelRef.value) {
+			labelWidth.value = labelRef.value.offsetWidth + 64
+		}
+	})
+
+	useSySelectAria({
+		textInput,
+		isOpen,
+		uniqueMenuId,
+		activeDescendantId,
+		isRequired,
+		hasError,
+		selectedItem,
+	})
+
+	const initializeActivatorProps = (activatorProps: Record<string, unknown>) => {
+		const asRecord = activatorProps as Record<string, unknown>
+		return {
+			...asRecord,
+			onKeydown: undefined,
+			onClick: undefined,
+			ref: (el: unknown) => {
+				textInput.value = el as InstanceType<typeof VTextField>
+				const refFn = (activatorProps as { ref?: (el: unknown) => void }).ref
+				refFn?.(el)
+			},
+		}
+	}
+
+	return {
+		isOpen,
+		labelWidth,
+		inputId,
+		uniqueMenuId,
+		formattedItems: formattedItems as unknown as ComputedRef<SySelectSetupItemType[]>,
+		hasError,
+		isRequired,
+		showHelpTextAsMessage,
+		showHelpTextBelow,
+		validateOnSubmit,
+		labelWithAsterisk,
+		calculatedWidth,
+		menuTarget,
+		labelRef,
+		htmlItemRefs,
+		textInput,
+		list,
+		toggleMenu,
+		closeList,
+		activeDescendantId,
+		onFieldKeydown,
+		onListKeydown,
+		isDefaultOption,
+		isItemSelected,
+		selectItem,
+		removeChip,
+		getItemText,
+		getChipText,
+		hasChips,
+		selectedChipsItems: selectedChipsItems as unknown as ComputedRef<Array<SySelectSetupItemType | string | number>>,
+		hasSelectionToClear,
+		selectedItemText,
+		initializeActivatorProps,
+	}
+}
